feat(user-service): adjust sign in response for MFA requirement
This commit modifies the sign in endpoint handler to return "202 Accepted" instead of "201 Created" when MFA is required to finish a sign-in process.

This distinction clearly indicates to the client that further action (MFA verification) is needed to complete the authentication process, even though initial credentials were valid and makes our API play nicer with common conventions, so it's all good!